Switch libvirt get_cpu_info method over to use config APIs
The get_cpu_info method in the libvirt driver currently uses
XPath queries to extract information from the capabilities
XML document. Switch this over to use the new config class
LibvirtConfigCaps. Also provide a test case to validate
the data being returned.

The latest fix for the review by Pushkar

The fix for Typo is done.

Change-Id: Icbbab9a8f3152c4961de325f5245542b8cbd839d
Closes-Bug: 1234567